Fix issue where applet fails to notice any updates after initial check.
In particular, g_timeout_add is expecting a pointer to a function which returns
a gboolean. Previously various implementations of the function for the different
backends were returning void. These now return true (a return value of false
indicates that the function should no longer be called). Issue spotted on x86_64
fedora 25.
Also removed dead code to do with aptget